New fonts, redesign, api level changed from 21 to 28 (To support font weight attribute)

+++ Qytap/app/src/main/java/com/tamertokbaev/qytap/models/Book.kt
@@ -2,7 +2,7 @@ package com.tamertokbaev.qytap.models
 
 data class Book(
     val author: String?,
-    val book_depository_stars: Double?,
+    val book_depository_stars: Float?,
     val category: String?,
     val created_at: String?,
     val currency: String?,
